### 2022-06-13 #54 @etam4260
hudpy: A Python interface for the US Department of Housing and Urban Development APIs
Submitting Author: @etam4260
Package Name: hudpy
One-Line Description of Package: A Python Interface to the US Department of Housing and Urban Development APIs
Repository Link:  https://github.com/etam4260/hudpy
Version submitted:  0.2.0
Version accepted: TBD

---

Currently the R version is under peer review at ROpenSci:
https://github.com/ropensci/software-review/issues/524

This package is a mapped over version from R to Python.

This package gives users access to the US Department of Housing and Urban Development APIs
available at HUD User:
https://www.huduser.gov/hudapi/public/login

This also has additional wrappers onto of the data provided from these APIs to determine if US geographies intersect
based on addresses and (cross walking a US dataset) explained in this paper below:

Wilson, Ron and Din, Alexander, 2018. “Understanding and Enhancing the U.S.
        Department of Housing and Urban Development’s ZIP Code Crosswalk Files,”
        Cityscape: A Journal of Policy Development and Research, Volume 20 Number 2, 277
        https://www.huduser.gov/portal/periodicals/cityscpe/vol20num2/ch16.pdf

> *Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.*

NA

It is a data retrieval package because it retrieves data from an API. It is data munging package because it implements the concept of cross walking a file (converting a dataset described in one US geographic identifier into another) explained in the paper mentioned above.

  -
I am hoping to reach professors, researchers, and students with this package. This gives access to the crosswalk files which is a geospatial technique described very well in these journal articles:

Din, Alexander and Wilson, Ron, 2020. “Crosswalking ZIP Codes to Census Geographies: Geoprocessing the U.S. Department of Housing & Urban Development’s ZIP Code Crosswalk Files,” Cityscape: A Journal of Policy Development and Research, Volume 22, Number 1, https://www.huduser.gov/portal/periodicals/cityscpe/vol22num1/ch12.pdf

Wilson, Ron and Din, Alexander, 2018. “Understanding and Enhancing the U.S. Department of Housing and Urban Development’s ZIP Code Crosswalk Files,” Cityscape: A Journal of Policy Development and Research, Volume 20 Number 2, 277 – 294.

Not any that I know of.

NA

- [X] has an [OSI approved license][OsiApprovedLicense].

In terms of vignette, I am assuming you are referring to a Jupyter notebook style documentation. I don't have this, but
there is a website with vignette like documentation: https://etam4260.github.io/hudpy/build/html/index.html

- [X] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

I feel like there should be a section for details that might not fit into any of the questions above.
Currently my package does not implement caching, but the skeleton code is currently in the module for future
implementation. It will work similar to the caching implemented in the rhud package currently being reviewed by ROpenSci.

I have not looked into the PEP 8 style guide yet, but will make sure to review that.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
