### 2022-06-14 #54 @NickleDave
Hi @etam4260 and welcome to pyOpenSci!

This definitely looks in scope at first glance.
Thank you for filling out the submission in detail.

Let me check in with everyone and we will get back to you by early next week.
