### 2022-06-28 #54 @NickleDave
Hi again @etam4260 sorry for the delayed reply.
Officially, pyOpenSci is on hiatus until fall while we move to  a new "home".
@lwasser is about to update the templates and README to make that clear.

But I would really like to get you a review.
I am sure `hudpy` is in scope, and it would be great to help you make sure it's as good as it can be for the Python community.

Please reply confirming that you would still like to initiate the review process.
Once you do so, I will start finding reviewers.
