### 2022-07-09 #54 @etam4260
Hi @NickleDave, thanks for the response. I've been taking a break from OSS development for a bit. Since I have submitted the rhud package for review, I am still working on changes for those and have not "mapped" over the feedback to the python version. I first would like to work on making those changes before getting formal feedback as I think the reviewers' comments will have a lot of overlap with reviews given so far in the R version. I think in two weeks should be fine. Thanks, again!
