### 2022-09-01 #54 @lwasser
hey there @etam4260 ! I really appreciate the conversation above and the acknowledgement that you needed a break! I wanted to check back in to see if you are still interested in working through a review with us or if you'd like to wait. IF you would like to wait - no problem. I will close this issue for now, and you can reopen it when you are ready. If you are ready now please just let us know.

Either way we are so happy to have you here at pyOpenSci - just let us know what timeing works best for you. If i don't hear from you within the next week, I will go ahead and close the issue. You can still reopen it at any time.

@NickleDave thank you again for serving as editor here!! 💥
