### 2022-09-01 #54 @lwasser
ok no problem @etam4260 ! I am going to close this issue with this comment. But please do feel free to resubmit in the future if you find that you have more time to devote to the review process. And keep an eye on pyopensci as we grow the organization and the community. i'm hopeful that we can really help the community of maintainers like you! All the best.
